Removed unnessary globals and refactored
Removed a creds file location.
Refactored the table names into a file so that it could be changed without changing the code. (This will later be stored on S3)
Changed prog_start to be used for headers for each code section.
Refactored the reading from MySQL and writing to S3/Disk so that it used an array of table names rather than hardcoding each one in.